requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.20
python-dotenv>=1.0.1
fastapi>=0.110.0
uvicorn>=0.30.0
